# docs/api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: 乡声集盒 Entry / Recording API
  version: 1.0.0
  description: |
    当前公开领域契约。Entry 是同一个词及其读音身份，EntrySense 是编号义，
    Recording 是一次录音；二者通过带角色的 RecordingEntryLink 多对多关联。
    已退役的 Can、Nameplate、Flavor、Package 与自动选主接口不属于本契约。
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
...
security:
  - bearerAuth: []
paths:
  /entries/:
    get:
      tags: [Entries]
      operationId: listEntries
      summary: 按词条优先查询
      security: []
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: dialect_id, in: query, schema: { type: integer } }
        - { name: dialect_scope, in: query, schema: { type: string, enum: [exact, subtree] } }
        - { name: has_recording, in: query, schema: { type: boolean } }
        - { name: ipa, in: query, schema: { type: string } }
        - { name: romanization, in: query, schema: { type: string } }
      responses:
        "200": { description: 词条列表, content: { application/json: { schema: { $ref: "#/components/schemas/EntryPage" } } } }
    post:
      tags: [Entries]
      operationId: createEntry
      summary: 创建词条初稿
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/EntryV2" } } }
      responses:
        "201": { description: 已创建, content: { application/json: { schema: { $ref: "#/components/schemas/EntryV2" } } } }
...
  /entries/{entry_id}/:
    parameters:
      - { $ref: "#/components/parameters/EntryId" }
    get:
      tags: [Entries]
      operationId: getEntry
      summary: 获取词条详情
      security: []
      responses:
        "200": { description: 词条详情, content: { application/json: { schema: { $ref: "#/components/schemas/EntryV2" } } } }
    patch:
      tags: [Entries]
      operationId: updateEntry
      summary: 修订词条
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/EntryV2" } } }
      responses:
        "200": { description: 已更新, content: { application/json: { schema: { $ref: "#/components/schemas/EntryV2" } } } }
  /entries/{entry_id}/bookmark/:
    parameters:
      - { $ref: "#/components/parameters/EntryId" }
    put:
      tags: [Entries]
      operationId: bookmarkEntry
      summary: 收藏词条
      responses:
        "200": { description: 收藏状态, content: { application/json: { schema: { $ref: "#/components/schemas/BookmarkResult" } } } }
    delete:
      tags: [Entries]
      operationId: unbookmarkEntry
      summary: 取消收藏词条
      responses:
        "200": { description: 收藏状态, content: { application/json: { schema: { $ref: "#/components/schemas/BookmarkResult" } } } }
  /entry-senses/:
    get:
      tags: [Entries]
      operationId: listEntrySenses
      summary: 查询编号义
      security: []
      responses:
        "200": { description: 编号义列表, content: { application/json: { schema: { $ref: "#/components/schemas/EntrySensePage" } } } }
    post:
      tags: [Entries]
      operationId: createEntrySense
      summary: 补充编号义
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/EntrySenseV2" } } }
      responses:
        "201": { description: 已创建, content: { application/json: { schema: { $ref: "#/components/schemas/EntrySenseV2" } } } }
...
  /pronunciation-variants/:
    get:
      tags: [Entries]
      operationId: listPronunciationVariants
      summary: 查询地区读音
      security: []
      responses:
        "200": { description: 地区读音列表, content: { application/json: { schema: { $ref: "#/components/schemas/PronunciationVariantPage" } } } }
    post:
      tags: [Entries]
      operationId: createPronunciationVariant
      summary: 补充地区读音
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/PronunciationVariantV2" } } }
      responses:
        "201": { description: 已创建, content: { application/json: { schema: { $ref: "#/components/schemas/PronunciationVariantV2" } } } }
...
  /recordings/:
    get:
      tags: [Recordings]
      operationId: listRecordings
      summary: 浏览录音流
      security: []
      responses:
        "200": { description: 录音列表, content: { application/json: { schema: { $ref: "#/components/schemas/RecordingPage" } } } }
    post:
      tags: [Recordings]
      operationId: createRecording
      summary: 提交录音及原始大意
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RecordingV2" } } }
      responses:
        "201": { description: 已创建, content: { application/json: { schema: { $ref: "#/components/schemas/RecordingV2" } } } }
...
  /recording-entry-links/:
    get:
      tags: [Recordings]
      operationId: listRecordingEntryLinks
      summary: 查询录音与词条关系
      security: []
      responses:
        "200": { description: 关系列表, content: { application/json: { schema: { $ref: "#/components/schemas/RecordingEntryLinkPage" } } } }
    post:
      tags: [Recordings]
      operationId: createRecordingEntryLink
      summary: 提出录音与词条关系
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RecordingEntryLinkV2" } } }
      responses:
        "201": { description: 已创建, content: { application/json: { schema: { $ref: "#/components/schemas/RecordingEntryLinkV2" } } } }
...
  /usage-attestations/:
    get:
      tags: [Entries]
      operationId: listUsageAttestations
      summary: 查询地区使用补证
      security: []
      responses:
        "200": { description: 补证列表, content: { application/json: { schema: { $ref: "#/components/schemas/UsageAttestationPage" } } } }
    post:
      tags: [Entries]
      operationId: createUsageAttestation
      summary: 确认我这里也这么说
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UsageAttestationV2" } } }
      responses:
        "201": { description: 已创建, content: { application/json: { schema: { $ref: "#/components/schemas/UsageAttestationV2" } } } }
  /usage-attestations/{attestation_id}/:
    parameters:
      - { $ref: "#/components/parameters/AttestationId" }
    get:
      tags: [Entries]
      operationId: getUsageAttestation
      summary: 获取地区使用补证
      security: []
      responses:
        "200": { description: 补证详情, content: { application/json: { schema: { $ref: "#/components/schemas/UsageAttestationV2" } } } }
    delete:
      tags: [Entries]
      operationId: retractUsageAttestation
      summary: 撤回本人的地区使用补证
      responses:
        "204": { description: 已撤回 }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    EntryId: { name: entry_id, in: path, required: true, schema: { type: integer } }
...
    AttestationId: { name: attestation_id, in: path, required: true, schema: { type: integer } }
...
  schemas:
    DialectRef:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: string }
        qualified_code: { type: string, description: 仅供 API、历史映射和管理工具使用 }
        path_names: { type: array, items: { type: string } }
        sort_order: { type: integer }
    UserRef:
      type: object
      nullable: true
      properties:
        id: { type: integer, nullable: true }
        username: { type: string }
        nickname: { type: string }
    EntryV2:
      type: object
      properties:
        id: { type: integer, readOnly: true }
        display_writing: { type: string, readOnly: true }
        summary: { type: string }
        identity_note: { type: string }
        usage_dialect: { $ref: "#/components/schemas/DialectRef" }
        status: { type: string, enum: [draft, reviewed, disputed, rejected] }
        visibility: { type: boolean }
        recording_count: { type: integer, readOnly: true }
        needs_audio: { type: boolean, readOnly: true }
        is_bookmarked: { type: boolean, readOnly: true }
        senses: { type: array, items: { $ref: "#/components/schemas/EntrySenseV2" }, readOnly: true }
        pronunciation_variants: { type: array, items: { $ref: "#/components/schemas/PronunciationVariantV2" }, readOnly: true }
        recording_links: { type: array, items: { $ref: "#/components/schemas/RecordingEntryLinkV2" }, readOnly: true }
    EntrySenseV2:
      type: object
      properties:
        id: { type: integer, readOnly: true }
        entry: { type: integer }
        sense_number: { type: integer }
        gloss: { type: string }
        usage_note: { type: string }
        status: { type: string, enum: [draft, reviewed, disputed, rejected] }
        concepts: { type: array, items: { type: object }, readOnly: true }
    PronunciationVariantV2:
      type: object
      properties:
        id: { type: integer, readOnly: true }
        entry: { type: integer }
        dialect: { $ref: "#/components/schemas/DialectRef" }
        ipa: { type: string }
        base_romanization: { type: string }
        surface_romanization: { type: string }
        reading_type: { type: string }
        status: { type: string, enum: [draft, reviewed, disputed, rejected] }
    RecordingV2:
      type: object
      properties:
        id: { type: integer, readOnly: true }
        audio_url: { type: string, format: uri }
        usage_dialect: { $ref: "#/components/schemas/DialectRef" }
        original_gloss: { type: string }
        recording_type: { type: string, enum: [word, phrase, sentence, narrative] }
        status: { type: string, enum: [draft, published, disputed, rejected] }
        visibility: { type: boolean }
        entry_links: { type: array, items: { $ref: "#/components/schemas/RecordingEntryLinkV2" }, readOnly: true }
    RecordingEntryLinkV2:
      type: object
      properties:
        id: { type: integer, readOnly: true }
        recording: { type: integer }
        entry: { type: integer }
        role: { type: string, enum: [primary, mention, competing] }
        status: { type: string, enum: [suggested, accepted, disputed, rejected] }
        is_current: { type: boolean }
    UsageAttestationV2:
      type: object
      properties:
        id: { type: integer, readOnly: true }
        entry: { type: integer }
        dialect: { $ref: "#/components/schemas/DialectRef" }
        attester: { $ref: "#/components/schemas/UserRef" }
        active: { type: boolean }
...
    BookmarkResult:
      type: object
      properties:
        entry_id: { type: integer }
        bookmarked: { type: boolean }
        changed: { type: boolean }
    Page:
      type: object
      properties:
        count: { type: integer }
        next: { type: string, format: uri, nullable: true }
        previous: { type: string, format: uri, nullable: true }
    EntryPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            results: { type: array, items: { $ref: "#/components/schemas/EntryV2" } }
    EntrySensePage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            results: { type: array, items: { $ref: "#/components/schemas/EntrySenseV2" } }
    PronunciationVariantPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            results: { type: array, items: { $ref: "#/components/schemas/PronunciationVariantV2" } }
    RecordingPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            results: { type: array, items: { $ref: "#/components/schemas/RecordingV2" } }
    RecordingEntryLinkPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            results: { type: array, items: { $ref: "#/components/schemas/RecordingEntryLinkV2" } }
    UsageAttestationPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            results: { type: array, items: { $ref: "#/components/schemas/UsageAttestationV2" } }
...